Changed data type of appVersion

+++ lib/src/main/java/se/gustavkarlsson/nag/Record.kt
@@ -4,6 +4,6 @@ data class Record(
 	val id: Long,
 	val key: String,
 	val timestamp: Long,
-	val appVersion: Int,
+	val appVersion: Long,
 	val value: String
 )
